backend/apimonitor/requirements.txt
Django==5.0.1
djangorestframework==3.16.1
celery==5.5.3
redis==5.0.1
httpx==0.25.2
scikit-learn==1.5.0
python-dotenv==1.1.1
environs==11.0.0
pytest==7.4.4
pytest-django==4.11.1
model-bakery==1.17.0
openai==1.6.1
psycopg2-binary==2.9.10
gunicorn==23.0.0